Fold NaquadahCore into NaquadahBrowser
This marks the beginning of a major change in the project. Naquadah has been completely rewritten. It is still lacking a bit of the functionality it had before such as event handling and style-class managment but the project is better designed so that it can move forward on firm footing ...hopefully.